Add arm and arm64 builds of clang-8 to production
Use the production build and boot infrastructure to build clang-8 for
a subset of architectures (armv7 and arm64).

As arm64 is a little noisy, and armv7 will likely be extremely noisy,
use a limited recipient list to not annoy the linux-next list.

As the arch-complete job can't currently deal with different email
recipients for different builds, use a different tree name until
such time as it can.